initial commit for nsf-template
made a skeleton proposal with relevant instructions from the NSF grant
proposal guide (GPG).  i split each section into a different tex file
for easier editing.

the formatting is pretty simple; there's no real need to define a latex
class for it.  the main thing this offers is to redefine the page
numbering for each section.  makes it easier for reviewers to know you
stayed within the page limits.

to upload a proposal to fastlane, you need to split each section into a
different file.  in order to do that automatically, i write the starting
and ending pages for each section to the logfile.  the ruby script
`split.rb' then parses this and runs ghostscript to do the splitting.  i
defined a special counter for this and use the everyshipout package to
update it.  this may be hacky and fragile, but it's not the end of the
world if it screws up...